refactor(git): move deprecated functions
